move cli tests to separate package
