+++ tests/cli-tests/fixtures/basic-project/schema.graphql
