+++ tests/cli-tests/fixtures/basic-with-no-ui.ts
