+++ tests/cli-tests/fixtures/custom-prisma-project/schema.prisma
