+++ tests/cli-tests/fixtures/log-node-env.ts
