+++ tests/cli-tests/fixtures/no-fields-with-migrations.ts
