+++ tests/cli-tests/fixtures/no-fields.ts
