+++ tests/cli-tests/fixtures/one-field-with-migrations.ts
