+++ tests/cli-tests/fixtures/two-fields-with-migrations.ts
