+++ tests/cli-tests/fixtures/with-ts.ts
